+++ lib/routes.dart
@@ -13,6 +13,7 @@ import 'pages/find_phone_page.dart';
 import 'pages/flash_script_page.dart';
 import 'pages/home_page.dart';
 import 'pages/instructions_page.dart';
+import 'pages/join_beta_page.dart';
 import 'pages/not_found_page.dart';
 import 'pages/privacy_policy_page.dart';
 import 'pages/recoveries_page.dart';
@@ -41,6 +42,7 @@ class AppRoutes {
   static const String community = '/community';
   static const String about = '/about';
   static const String privacy = '/privacy';
+  static const String joinBeta = '/join-beta';
 
   static String romDetail(String id) => '/roms/$id';
   static String recoveryDetail(String id) => '/recoveries/$id';
@@ -234,6 +236,11 @@ GoRouter buildRouter() {
                 builder:
                     (BuildContext _, GoRouterState __) =>
                         const PrivacyPolicyPage(),
+              ),
+              GoRoute(
+                path: 'join-beta',
+                builder:
+                    (BuildContext _, GoRouterState __) => const JoinBetaPage(),
               ),
             ],
           ),